requirements.txt
tkinter
Pillow
nltk
gensim
